Move Optimization functionality from LCP class to a new LCP_Optimizer class
